+++ packages/examples/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 All notable changes to this npm module are documented in this file.
 
-## [2024.10.5] - 2024-10-2x
+## [2024.10.5] - 2024-10-30
 
+- Updated to `monaco-languageclient@9.0.0-next.6`, `monaco-editor-wrapper@6.0.0-next.6` and `@typefox/monaco-editor-react@6.0.0-next.6`.
+- Updated all `@codingame/monaco-vscode` packages to `10.1.3`.
 - Added clangd example.
 - Added application playground example featuring the views service override.
 
@@ -15,7 +17,7 @@ Updated all `@codingame/monaco-vscode` packages to `10.1.1`.
 ## [2024.10.3] - 2024-10-21
 
 - Updated to `monaco-languageclient@9.0.0-next.4`, `monaco-editor-wrapper@6.0.0-next.4` and `@typefox/monaco-editor-react@6.0.0-next.4`.
-Updated all `@codingame/monaco-vscode` packages to `10.1.`.
+Updated all `@codingame/monaco-vscode` packages to `10.1.0`.
 - Updated to eslint 9
 
 ## [2024.10.2] - 20241-10-11